@@ -4,54 +4,55 @@
     "private": true,
     "dependencies": {
         "@feathersjs/authentication": "^4.5.11",
-        "@feathersjs/authentication-client": "^4.5.7",
+        "@feathersjs/authentication-client": "^4.5.11",
         "@feathersjs/client": "^4.5.11",
-        "@feathersjs/feathers": "^4.5.7",
+        "@feathersjs/feathers": "^4.5.11",
         "@feathersjs/rest-client": "^4.5.11",
-        "@material-ui/core": "^4.11.2",
+        "@material-ui/core": "^4.12.3",
         "@material-ui/icons": "^4.11.2",
-        "@reduxjs/toolkit": "^1.5.1",
-        "@testing-library/react": "^10.4.9",
-        "@testing-library/user-event": "^12.8.3",
-        "@types/lodash": "^4.14.168",
-        "@types/mocha": "^8.2.3",
+        "@reduxjs/toolkit": "^1.6.2",
+        "@testing-library/react": "^12.1.2",
+        "@testing-library/user-event": "^13.5.0",
+        "@types/lodash": "^4.14.177",
+        "@types/mocha": "^9.0.0",
         "@types/moment-timezone": "^0.5.30",
-        "@types/node": "^14.14.44",
-        "@types/react": "^16.14.6",
-        "@types/react-dom": "^16.9.12",
-        "@types/react-redux": "^7.1.16",
-        "@types/react-router-dom": "^5.1.7",
-        "@types/react-typist": "^2.0.1",
-        "acorn": "^7.3.1",
-        "axios": "^0.21.4",
+        "@types/node": "^16.11.10",
+        "@types/react": "^17.0.37",
+        "@types/react-dom": "^17.0.11",
+        "@types/react-redux": "^7.1.20",
+        "@types/react-router-dom": "^5.3.2",
+        "@types/react-typist": "^2.0.3",
+        "acorn": "^8.6.0",
+        "axios": "^0.24.0",
         "bezier-easing": "^2.1.0",
-        "dotenv": "^8.2.0",
-        "framer-motion": "^2.1.4",
+        "dotenv": "^10.0.0",
+        "framer-motion": "^4.1.17",
         "hex-to-rgba": "^2.0.1",
-        "immutable": "^4.0.0-rc.12",
-        "inversify": "^5.0.1",
+        "immutable": "^4.0.0",
+        "inversify": "^6.0.1",
         "lodash": "^4.17.21",
         "map-range": "^0.1.2",
         "minimalist": "^1.0.0",
         "minimist": "^1.2.5",
-        "react": "^16.14.0",
-        "react-awesome-reveal": "^2.4.2",
-        "react-cookie": "^4.0.3",
-        "react-dom": "^16.14.0",
+        "react": "^17.0.2",
+        "react-awesome-reveal": "^3.8.1",
+        "react-cookie": "^4.1.1",
+        "react-dom": "^17.0.2",
         "react-flicker": "^1.1.1",
-        "react-live-clock": "^4.0.5",
-        "react-particles-js": "^3.3.0",
-        "react-redux": "^7.2.4",
-        "react-router-dom": "^5.2.0",
-        "react-scripts": "3.4.1",
-        "react-spring": "^8.0.27",
-        "react-svg": "^11.0.32",
+        "react-live-clock": "^5.3.1",
+        "react-moment": "^1.1.1",
+        "react-particles-js": "^3.6.0",
+        "react-redux": "^7.2.6",
+        "react-router-dom": "^6.0.2",
+        "react-scripts": "4.0.3",
+        "react-spring": "^9.3.2",
+        "react-svg": "^14.1.3",
         "react-svg-hexagon": "^1.0.0",
         "react-transition-group": "^4.4.2",
         "react-typist": "^2.0.5",
         "reflect-metadata": "^0.1.13",
-        "styled-components": "^5.1.1",
-        "typescript": "~4.2.3",
+        "styled-components": "^5.3.3",
+        "typescript": "~4.5.2",
         "winston": "^3.3.3"
     },
     "scripts": {
