@@ -1,25 +1,23 @@
 import React from "react";
-import { Switch, Route } from "react-router-dom";
-import Routes from "../app/constants/routes";
+import { Routes, Route } from "react-router-dom";
+import { Routes as LocalRoutes } from "../app/constants/routes";
 import DashboardLanding from "../pages/dashboard/DashboardLanding";
 import DashboardLendee from "../pages/dashboard/dashboard-lendee/DashboardLendee";
 import DashboardLender from "../pages/dashboard/lender/DashboardLender";
 
 const DashboardRouter: React.FunctionComponent = () => {
     return (
-        <Switch>
-            <Route
-                exact
-                path={Routes.LENDEE_DASHBOARD}
-                component={DashboardLendee}
-            />
-            <Route
-                exact
-                path={Routes.LENDER_DASHBOARD}
-                component={DashboardLender}
-            />
-            <Route path={Routes.DASHBOARD} component={DashboardLanding} />
-        </Switch>
+        <Routes>
+            <Route path={LocalRoutes.LENDEE_DASHBOARD}>
+                <DashboardLendee />
+            </Route>
+            <Route path={LocalRoutes.LENDER_DASHBOARD}>
+                <DashboardLender />
+            </Route>
+            <Route path={LocalRoutes.DASHBOARD}>
+                <DashboardLanding />
+            </Route>
+        </Routes>
     );
 };
 
