@@ -27,7 +27,7 @@ export const { setFinished } = slice.actions;
 // The function below is called a selector and allows us to select a value from
 // the state. Selectors can also be defined inline where they"re used instead of
 // in the slice file. For example: `useSelector((state: RootState) => state.counter.value)`
-export const selectGreeting = (state: RootState) => {
+export const selectGreeting = (state: RootState): boolean => {
     return state.greeting.finished;
 };
 
