@@ -1,5 +1,3 @@
-
-
-export const handleConnectionUpdate = (e: any) => {
+export const handleConnectionUpdate = (e: unknown): void => {
     console.log("connection changed:", e);
-}
+};
