@@ -14,7 +14,13 @@ export type Range = [number, number];
  * @param x2 start point for target range
  * @param y2 end point for target range
  */
-export const mapValue = (x1, y1, x2, y2, value): number => {
+export const mapValue = (
+    x1: number,
+    y1: number,
+    x2: number,
+    y2: number,
+    value: number
+): number => {
     return MapRange(linear, x1, y1, x2, y2)(value);
 };
 
@@ -35,7 +41,9 @@ export const mapValues = (values: number[], range?: Range): number[] => {
     });
 };
 
-export const useMapper = (values: number[], range?: Range): Function => {
+// Disable inspection because this return type is a pain.. and I don't want to deal with it yet..
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export const useMapper = (values: number[], range?: Range): any => {
     const maxValue = Math.max.apply(null, values);
     const minValue = Math.min.apply(null, values);
     let targetStart, targetEnd;
